flax_greenhouse_mvc/requirements.txt
matplotlib>=3.5.0
numpy>=1.20.0
paho-mqtt>=1.5.0
